refactor: reorganize utils package with date-time and result modules
- Remove unused @polkadot/util dependency
- Create date-time.ts module with consolidated timestamp functions
- Reorganize result types into result/ directory structure:
  - result/sync.ts: Core synchronous result functionality
  - result/async.ts: Async result operations
  - result/index.ts: Main export interface
- Update all import paths to use new module structure
- Consolidate duplicate timestamp functions into single nowISOString()
