+++ packages/utils/src/index.ts
@@ -1,4 +1,3 @@
-import { DateTime } from "luxon";
 import { assert, AssertionError } from "tsafe";
 
 export * from "./typing.js";
@@ -64,41 +63,7 @@ export function bigintDivision(a: bigint, b: bigint, precision = 8n): number {
 
 // == Time ==
 
-export function getExpirationTime(
-  blockNumber: number | undefined,
-  expirationBlock: number,
-  relative = false,
-): string {
-  if (!blockNumber) return "Unknown";
-
-  const blocksRemaining = expirationBlock - blockNumber;
-  const secondsRemaining = blocksRemaining * 8; // 8 seconds per block
-
-  const expirationDate = DateTime.now().plus({ seconds: secondsRemaining });
-  if (relative) {
-    return expirationDate.toRelative();
-  }
-  return expirationDate.toLocaleString(DateTime.DATETIME_SHORT);
-}
-
-export function getCreationTime(
-  blockNumber: number | undefined,
-  creationBlock: number,
-  relative = false,
-) {
-  if (!blockNumber) return "Unknown";
-
-  const blocksAgo = blockNumber - creationBlock;
-  const secondsPassed = blocksAgo * 8; // 8 seconds per block
-
-  const creationDate = DateTime.now().minus({ seconds: secondsPassed });
-
-  if (relative) {
-    return creationDate.toRelative();
-  }
-
-  return creationDate.toLocaleString(DateTime.DATETIME_SHORT);
-}
+export { getExpirationTime, getCreationTime } from "./date-time.js";
 
 // == String ==
