+++ packages/utils/src/result/index.ts
@@ -1,0 +1,2 @@
+export * from "./sync.js";
+export * from "./async.js";
